Add smooth_spiralized_contours setting to control smoothing of contours when spiralizing.
Smoothing the contours minimises the visibility of the z-seam and for most models is
probably a good idea. Where it is less good is when the model has a fine surface details
which will get mangled by the smoothing. So we should let the user decide whether they
want to smooth or not. The default is true as most of the time it's a good thing.

# Conflicts:
#	resources/definitions/fdmprinter.def.json