WARNING: TO BE REMOVED! impossible settings limit_to_extruder
CURA-3740

These settings inherently have to do with multiple features and
so we should find a value function which depends on
settings from the extruder trains corresponding to those features (wall, skin, infill).

However, we haven't found those functions yet and we have to do research.

Once we have found the formula these settings shouldn't depend on a single extruder any more
and so we should remove the limit_to_extruder.