Make extruder_nr settable per extruder
It is actually being set per extruder right now, since this is where the definition gets its extruder number from. If the setting is not settable per extruder, the ExtruderStack skips the setting entirely and defers to the global stack, which doesn't have the setting and so None is returned. This was giving errors.

Contributes to issue CURA-3772.