Resolve bed adhesion type via limit-to-extruder
The cause for this change is that TPU says we want to print with brim but CPE wants to print with raft. In those cases the resolve function would choose the raft because that's safer. However if the default extruder happens to be the TPU one you then get a raft made of TPU which is flexible. That defeats the purpose of the raft, and causes the raft to warp, clog on the nozzle, destroy the world, etc. So there, I saved the world once more.