Order quality results by "supported" first
CURA-3884

When we create a new material, that meterial's "metadata/material"
will be set to "Custom", which in other cases are normally "PLA",
"ABS", etc. This becomes a problem when we try to get quality
profiles for this custom material: both supported and not-supported
qualities will be returned. In this case, we order the result list
by supported first, so Cura will pick a supported quality for this
material if possible.