Limit Raft Interface Thickness to a minimum of 0.225mm
If it's less, the layer doesn't get properly filled, according to the materials team. They requested that the thickness should be set to 0.225mm for just the Fine and Extra Fine profiles, but I opted for a formula that would still let it scale with the layer height up to a point.