Fix MIME type of material resources
Because of this wrong MIME type, the material profile would get saved using the preferred extension of the instance container, .inst.cfg. This would then get interpreted as having the MIME type of instance containers, and then it would fail to load because it is not an INI file.

Now it stores it as .xml.fdm_material and so the MIME type database interprets it as an XML file.

Contributes to issue CURA-3937.