Remove Not-Supported profiles for CPE+ and PC
They are now supported. This breaks the stacks of people who had these profiles active when they upgraded from 2.6.0 to 2.6.1, which causes them to get a screen of which printer they want to add, and they lose their current settings. But we decided that this use case is too slim to make an entire version upgrade for this patch-version release.

Contributes to issue CURA-3960.