Optimise Ultimaker 2+ polypropylene profiles
These are optimised to the level of the materials: Anything that is common between quality profiles with the same material and nozzle gets bubbled to the profile for that material (per nozzle). It doesn't bubble any further (to the variant) because that could have an impact on the other profiles.
This basically only removes settings because where these settings then bubble to is the material profile in the fdm_materials repository. And a lot of settings are just removed because they are equal to the default.

Contributes to issue CURA-3960.