Increase warning value for outer wall wipe distance
For polypropylene a wipe distance of twice the nozzle size is fine. Even required, since the material warps so much that it breaks open automatically if a breach starts to appear at the seam. For other warping materials this could also be a problem, so let's just increase the warning value.

Contributes to issue CURA-3960.