Remove minimum warning value for z-hop height
It was giving a warning for 0.8mm polypropylene because that was set to 0.5mm hop. It was decided that a minimum value at all for z-hop is nonsense because you can just disable z-hop without getting a warning. I'm still setting it to zero because negative will still go wrong in your print.

Contributes to issue CURA-3960.