Remove gradual infill for PP
This was delivered false by the materials team. It seems that in the end they decided to not use gradual infill for PP but only made the change in UM3, not UM2+. Removing gradual infill also requires me to set the infill density to 20% (the default).

Contributes to issue CURA-3960.