Add 'Not Supported' profile for Polypropylene 0.25mm
Previously we also called the ID 'Not Supported' but we decided against that for new profiles, since that would mean that if the profile ever becomes supported we'd have to do a rename of the user's file names in their settings via a version upgrade. So now the ID is as if it's supported but the user-visible name still says that it's Not Supported and the profile still becomes orange.

Contributes to issue CURA-3960.