Simplify inheritance formula for prime tower wall thickness
Firstly, all the resolveOrValue calls are identity functions now so they can be phased out.

Secondly, all the max(extruderValues(...)) should be done by the resolve property rather than the value property. The resolve property already makes sure that the prime tower is evenly thick for every extruder, so it doesn't need to ensure this in the value function as well.

Contributes to issue CURA-3521.