Make sure the preferences file doesn't always get upgraded
When the setting is equal to the default the setting_version won't get written to the preferences file. In that case the version upgrade system assumes that the setting_version was 0 or 1 and upgrades it.

This makes sure that the setting is never equal to the default so that it always gets written.

Found while working on CURA-4024.