Temporary work around for "not overriden" limit to extruder features
CURA-4145

When features with limit to extruder set to "not overriden" and its
value is retrieved with "extruderValue()", because the extruder number
is "-1", extruderValue() falls back to the global stack. As a result,
values can be incorrect. This workaround using the variants temporarily
solves the problem for outer and inner wall line widths.