Make opacity of print monitor themeable
The new dark design specifies a colour for the print monitor overlay that is darker than what would be attainable with the default 75% opacity. So I'm making it 100% opacity (but still with a fade, so we need the opacity) and let the theme specify what opacity it wants.

Contributes to issue CURA-4148.