Switch to dark colours
This is the big one.
The progress bar buttons and progress bars haven't been properly updated yet, since I have no new information about those. Also, some things currently look kind of weird because they were designed for a different layout. For instance, the layer view slider has a black rails on a black background, since it was designed for being on the white background of the 3D view.

Contributes to issue CURA-4148.