Rename setting_validation_error/warning and border
In some places we want to use the bright error or warning colours to display that something is going on. In those places we'd like to refer to the colour setting_validation_error rather than setting_validation_error_border or something, so I'm renaming these four theme items.

Contributes to issue CURA-4148.