Add minimum_value to z_offset_taper_layers.
It now constrains the value to be no less than 0 when z_offset_layer_0 is negative or
z_offset_taper_layers is 0, otherwise the value must be such that the gap between subsequent
layers is not less than 80% of the normal layer height.

If the print head was lowered for the first layer, the tapering will always cause the
subsequent layers to be at least the normal layer height above the previous layer so the
tapering can occur over any number of layers.

On the other hand, if the print head was raised for the first layer, we demand that either
tapering is not used (value is 0) or that the tapering occurs over sufficient layers so that
the inter-layer gaps are not too small.