New base theme: Cura-light
There is a new theme now, Cura Light, based on the original colours of Cura 15.6 through 2.7. It is now also the base of the themes. It will soon be made the temporary default until the UX team can come up with something better for the dark theme.

Contributes to issue CURA-4148.