Let user accept/decline a plugin license if present
CURA-4222

If a downloaded plugin contains a license file, a dialog which contains
the license information will pop up. The user has to accept the license
before he/she can install the plugin.

If a plugin doesn't have a license file, after it is downloaded, it will
be installed directly just like before.