Revert making the camera image black while it doesn't react
This was giving segfaults sometimes on my computer. I suspect it's because the camera image could have been written to by both this Python code and by the camera itself, giving it a sort of data race or maybe that the image was discarded by the camera while it's being written to by Python. In any case, this should make it more stable.

Contributes to issue CURA-4376.