Add translations for strings that were added after string freeze
Please let's not do this again.

The Spanish, Dutch, French, German and Chinese translations were asked to a native or secondary speaker of the language in the office. The Finnish, Italian, Russian and Turkish translations are obtained from Google Translate (with some more clever searching than copy-paste). The Korean and Japanese translations are not needed since we don't translate cura.pot for those languages. The Polish and Brazillian Portuguese translations will be made after receiving the translations from the translators.

Contributes to issue CL-118.