Change error/success messages for checking cluster-capability
This breaks the string freeze. The developers of this plug-in are well aware. However, it turns out that for these two strings the string freeze had already broken. We'll change these strings first then translate them out of band. I hate this sort of menial administration that could've been prevented...

Contributes to issue CURA-4341.