Don't make filter_out_tiny_gaps a child setting of fill_perimeter_gaps
Parent settings are not used by the engine. Or they shouldn't be, because the parent settings are disabled by the front-end when all of their child settings are overwritten. Then you'd have a parent setting that is used by the engine but you can't modify it. Also, the command-line slicing doesn't load parent settings since it can't parse categories and such.