Let _fetchInstanceContainers return containers split by loaded or not
It must now return two dictionaries: One for the profiles that have been completely loaded and one for the profiles that are only metadata. We could probably improve on these a little bit, since all of these (except the material model) will now load all available quality profiles. I'll see if it is necessary to optimise that.

Contributes to issue CURA-4243.