Trust the stack values more than the decorator
CURA-4705

A SceneNode and its decorators can be deepcopied. However, the data in
some decorators will only be updated when a per-object settings stack
triggers a property changed event. That event cannot copied. So, it can
happen that a deepcopied SceneNode has inconsistent data in some of its
decorators than what's in the per-object settings stack.