Disable Traditional Chinese
We didn't get a new translation for 3.2. Sorry. People that had their preferences set to Traditional Chinese will stay on that language until they choose another, but the translation will be incomplete. If someone wants to pick up the translation for a next release, we'll keep the files in our source code.

Contributes to issue CURA-4883.