Add 512 padding bytes to the end of the g-code
Due to a bug in the firmware of the Ultimaker 2, the print was already ending before the last few commands were executed completely. It would then finish the last few commands in a different routine but the extrusion modifier wasn't applied there, resulting in a blob on your print.
This is a workaround that puts always one file system page (512 bytes) of padding at the end of the g-code file. If these comments are not executed that is fine. If they are then it's also fine. Until the bug is fixed in firmware, this will at least prevent the issue from occurring for users.