Don't make print estimates translateable
This was very very confusing for our translators. And all of the translators basically came down to the same translation: Leave it untranslated. So let's make it a bit easier on them and don't allow this to be translated.
If we really want the currency to be translated to the correct position, Qt has currency printing in its locale as an option too.