add Print Infill Every & Print Support Infill Every
To the user:
Change the setting Infill Layer Thickness to Print Infill Every (# layers), which seems to be more understandable. As the value of Infill Layer Thickness has to be multiple of the layer height it makes it easier to the user to set the setting.
Same for Support Infill Layer Thickness