Only display plug-in checker when there's really a warning
As discussed with LukeChen-Ultimaker, we only want to display a button when there is really something up with the current set-up, not when there might be something up. Most users will be confused to see a button that always says that everything is fine.

This also simplifies the signal handling a lot.

Contributes to issue CURA-4557.