Don't crash Cura when material file is wrongly formatted
That was the whole point of this change. Instead degrade gracefully. It won't load the material file but give an error message saying that the material is corrupt. Then it won't be able to load the stack as well because the material doesn't exist, and give an error about the stack as well.

Contributes to issue CURA-5045.