Simplify issue template
We don't usually need to know what specific Qt version or graphics driver the user has, and it's a great hassle for the user to figure this out, especially when they aren't technical and just installed an official release.
I hope that this leads more bug reporters to actually fill in the template.