Add setting for maximum travel resolution
I'm making this settable per extruder since a travel move is made with a certain active nozzle but can be between meshes so it can't be a per-mesh setting. I wouldn't know why you would want to set it differently per extruder though, but you can...

Contributes to issue CURA-5171.