Fix binding loop with height of tiles
The lining at the bottom of a tile was anchored to the bottom but part of the tile and part of its childrenRect. This caused a binding loop because the childrenRect is dependent on the positioning of this lining (and its height) and the positioning of the lining was dependent on the height of the tile.

Contributes to issue CURA-5035.