CURA-5357 Search for license files
The old code had the problem that by definition it skipped the `files/` folder and any other dir which of course is where the licenses are.

Now, so long as the license is always just called `LICENSE` (which is the convention in open source software), it can be located anywhere.