CURA-5280 Preserve manually entered project name
Manual override of job name should also set the base name so that when the printer prefix is updated, it the prefix can be added to the manually added name, not the old base name.

This was a bit tricky because if you enter a job name, is that the job name or the base name? My answer is that it's both, and if you update your printer for the job, it will append a new prefix.

If this is not the desired behavior, and in fact once a custom name is set nothing can change it, then line 78 should be removed.