Only load loading.gif when needed
Toolbox crashes on Mac OS X due to reaching the max limit of file
handlers. This seems to be caused by the loading the "loading.gif"
animation image. Probably because many widgets are created and each of
them has an animation image, and Qt (5.8?) seems to load everything at
the same time.