Add fan speed to basic setting visibility
It may not be what the users often want to change, but it is something that they may be expecting to see in the custom settings mode by default. It's something that you often think of as a typical setting but in reality you don't often want to change it. Let's get closer to what the user expects. Requested by the support team of Ultimaker.