Don't override boundingbox of SceneNode
Whatever was intended here, it didn't work. Because when a method is called it is called on the class of the object. Methods are not attributes of instances but attributes of the class. The attribute of the class was never changed, so this had no effect.

Contributes to issue CURA-5330.