Fix Z component of camera transformation
I discovered that it worked well when the camera was exactly horizontal and on the XY plane. Turns out that there was still something wrong with the Z component of the camera transformation.
Now the only problem I know of is when the viewport does not match the window size. We'll take a look at that.

Contributes to issue CURA-5482.