Construct support a bit before the surface
Because if it's in the surface then the support falls away due to the line width and due to the Z distance. It may still fall away due to the X/Y distance but I don't really have a solution for that yet. It's possible, I'm sure...

Contributes to issue CURA-5482.