Remove bottom-facing parts of UM3 platform
Since we have backface culling, it should appear correct when viewing from the top, but if you view from the bottom the build plate should now be gone completely so you can edit the scene more easily from the bottom. This is what you'd want for custom support.

Contributes to issue CURA-5482.