Don't construct support in a job
It's now so fast that it's not really necessary any more. By computing the support in the main thread, we prevent redrawing and so the red part won't disappear for a while while it is processing. It'll just stay there until processing is complete. This makes the support drawing feel much smoother.

Contributes to issue CURA-5482.