Remove prime_tower_wall_thickness setting
The wall thickness now depends on the order in which the prime towers are folded, which is really hard to obtain within an inheritance formula. So we forego one of the settings and compute the required thickness to obtain the desired volume in the engine where the shell order is known.

Contributes to issue CURA-5457.