Remove references to deleted settings
Also convert the values for prime_tower_min_volume to make it behave similar to what the minimum volume used to be. This feature fixed a bug with the computation from volume to thickness. Now it's correct but we have to correct the old values to be still the same (approximately) as what used to be the end result.

Contributes to issue CURA-5457.