Adjust maximum zoom to printer size
Previously the maximum distance you could zoom out was 2000. Now it's variable depending on how large your printer is. I've put it to 5 so that it works out to be approximately the same maximum zoom level for normal sized printers (like 25 cube build volume approximately).

This should make it possible to zoom out completely for large sized printers.