Add material_diameter into fdmprinter
When saving a material to a file, it will save all the settings in that
container, which needs to create SettingInstances for all the cached
ones, and for those instances, their definitions will be retrieved from
the machine definition.

material_diameter is one of the settings, but it only exists in the
extruder definitions, so when it tries to save a material profile, a lot
of warnings/errors will occur due to the missing "material_diameter" in
fdmprinter. Adding it back fixes this problem.